refactor(component): CursorOverlay 컴포넌트 분리 및 개선 / DP-136
